move requests utils
